Break Turnstile functionality into shared library
